Add frame_utils with external render func
