+++ setigen/__init__.py
@@ -11,3 +11,4 @@ from .split_utils import *
 from .stats import *
 from .time_freq_utils import *
 from .unit_utils import *
+from .frame_utils import *
