+++ setigen/frame_utils.py
@@ -1,0 +1,43 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+
+def db(x):
+    """
+    Converts to dB.
+    """
+    return 10 * np.log10(x)
+
+
+def render(data):
+    """
+    Display frame data in waterfall format.
+    """ 
+    plt.imshow(data,
+               aspect='auto',
+               interpolation='none')
+    plt.colorbar()
+    plt.xlabel('Frequency (px)')
+    plt.ylabel('Time (px)')
+    
+
+def integrate_frame(frame, normalize=False):
+    """
+    Integrate over time using mean (not sum).
+    """
+    data = frame.data
+    if normalize:
+        m, s = frame.get_noise_stats()
+        data = (data - m) / (s / frame.tchans**0.5)
+    return np.mean(data, axis=0)
+
+
+def integrate_frame_subdata(data, frame=None, normalize=False):
+    """
+    Integrate a chunk of data assuming frame statistics using mean (not sum).
+    """
+    if normalize:
+        assert frame is not None
+        m, s = frame.get_noise_stats()
+        data = (data - m) / (s / frame.tchans**0.5)
+    return np.mean(data, axis=0)
